CAMEL-13470: Publish security advisories
Adds two layouts for `security` and `security-advisory` and a archetype
for `security-advisory`. Security advisories are now served from
`/security` (instead of `/security-advisories.data`) so a `.htaccess`
file was added to redirect from previous URL to the current one.

README was updated to show how to use the Hugo archetype to publish a
new security advisory.