chore: remove critical
In order for the Content Security Policy be effective we either need to
place a random nonce generated per request to all inline script/styles
or determine the hashes of all inline script/styles and add them to the
Content-Security-Policy header.

Since the only inline script/style we have is put inline by critical its
easier to remove critical and rely on other optimization/minification
methods.