fix: conservatively collapse whitespace
When minifying the HTML non-conservative whitespace collapse would join
text from a paragraph with the tag, this in some cases would lead to the
joining of two words. For example, on the downloads page:

```html
<li>The latest feature release of Camel 2.24.x is <a href="../releases/release-2.24.1/">2.24.1</a></li>
```

Would be minimized as

```html
<li>The latest feature release of Camel 2.24.x is<a href="../releases/release-2.24.1/">2.24.1</a></li>
```

Joining the "is" with "2.24.1" and when rendering it would render as
"is2.24.1" where "2.24.1" would be rendered as a link.