chore: add HTML validation
Uses `html-validate`[1] to validate HTML for syntax, best practice and
accessibility.

Some of the checks needed to be disabled (via `.htmlvalidate.json`) as
the issues manifested in them would need to be addressed in Asciidoctor
conversion to HTML (no-inline-style, element-required-attributes,
prefer-tbody) and some are byproduct of minimization of HTML
(attr-quotes).

Also refactores the build, so now `documentation` and `website` tasks
are prefixed with `build:` and `checks` is used to run the
`xref-validator` and the `html-validate`.

[1] https://html-validate.org/