chore: upgrade Antora default UI to latest
This upgrades Antora default UI to the latest version and ports changes
made to have the Camel branding to it. Antora default UI has moved a lot
of the configuration into the variables and this should make it easier
to maintain the Camel Antora UI based on it.

Also removes the Serif font variants from use, somehow it didn't seem to
be a good fit and deals with some visual quirks like the last dropdown
menu being off screen.