chore: add cache busting
This adds cache busting by adding the hash of the resource to the built
file name, for example `site.css` becomes `site-41b4bf62e5.css`. This
allows us to cache the resources for long periods and still have them
loaded in the browsers at the latest versions when we do change them.

Ref: CAMEL-13816