chore: redesign and css consolidation
Redesigns the front page to use consistent type and adds a few whirls to
make the page look more interesting.

Also consolidates the CSS used between Antora and the Hugo built parts
very much a work in progress. The end result would be same exact styling
that we have in the Antora bits in the Hugo built bits and also
integration of only the used bits in Bootstrap into the CSS we have.

For this markup also needed to change and some visual differences will
be noticed between current and the proposed changes here.

A lot of the repeating layouts were removed so that the default layout
is used instead, thus reducing some of the repetition and easing future
maintenance.