chore: fix table responsiveness
This adds a Asciidoctor.js extension to post process the resulting HTML
to add a `<div>` wrapper around the `<table>` elements. The purpose of
this wrapper is to allow left-right scrolling of the table when the
table can't fit into the page width.
When the table can't fit into the page with the mobile clients usually
zoom-out to fit the full with and make the text much smaller and really
difficult to see.
This also refines the hyphenation, so that it's disabled only for the
first column (the column containing the component/endpoint option) and
enables it for other columns.