chore: remove inline styles in asciidoc content
With the content security policy set to disable any inline styles any
inline styles that are present are not taken into account. This adds a
Asciidoctor.js extension to remove all inline styles from the rendered
HTML to reduce the number of errors reported in the Google search
console and the browser console.