refactor: blog images and code widths
This adds and changes a number of featured images on blog posts. Some
featured images were replaced with royalty free images of higher
resolution, some are recreated as mash ups of two royalty free images.

Some blog posts didn't have a featured image, but included an image
after the first paragraph, making the blog post stand out against other
blog posts. The design dictates that a blog post starts with a featured
image (if present).

Design (css) was altered to make sure that no image is auto-resized
beyond 800px, the featured images and the shortcode (`{{ < image > }}`)
resize the image to 800px and the responsiveness of the image would take
up the whole width if another element would push the blog content wider.
This would make the aside content narrow and hard to read.

Font size of the code blocks slightly decreased and the `<pre>` tags on
blog posts are now word-wrapped so long lines in code samples don't
increase the width of the blog text content making the aside content too
narrow as a result.