fix: clean Jenkins workspace
Seems that a pull request generated some files in the workspace that are
now causing build failures. This will clean the workspace in a post
build step (always).

This should be removed if the build after the workspace is cleaned
succeeds, we cache node packages and git data in the worksapce, having
the workspace cleaned on every build would increase the build times.