fix: xref-validator with Antora 2.3-beta
The xref-validator needed to be updated to support the `start_paths`
configuration in the Antora playbook. Since we're using Yarn PnP
dependencies of xref-validator needed to be explicitly defined as the
versions that xref-validator depends on don't match the versions we're
using; the dependencies are to 2.3-alpha, and we're on 2.3-beta.