feat: change frontpage swoosh animation
To freshen up the frontpage a bit, and perhaps to be more representative
of the integration nature of Camel, this changes the swoosh animation to
a "transmission" animation. By using dash pattern of the same SVG paths
we can simulate "messages" traveling from left and right edges of the
screen.