chore: rebase on latest Default Antora UI
This rebases the changes we made on top of the latest state of the
Default Antora UI. I've tweaked the bottom margin of the TOC, so that
the items do not render beneath the explore panel (panel with the
components/versions). And I've removed scrollbar customizations, I find
them gimmicky and they're negatively impacting the usability.

The major new feature is the new page table of content menu on the right
of each page.