CAMEL-15337: Add TOC to static content
Based on work by @AemieJ in #446 this simplifies the implementation by
relying on the implementation already provided by Antora.

The main issue with reusing the Antora implementation was that the
markup of rendered Asciidoc contains `<div class="sect?">` elements
around the chapters. Hugo rendered, static, content does not have this,
so the logic for the _on this page_ functionality needed to be adjusted
slightly because of this.