fix: disable link-checker on a subset of files
Seems that the `link-checker` is reporting a false positive in
camel-kafka-connector component, the link[1] to
`/community/mailing-list/` is treated as relative within the component,
whereas is absolute towards to root of the website. Within the rendered
HTML the links is valid, the page `/community/mailing-list/index.html`
does exist.

This adds `--url-ignore` for `/community/mailing-list/` so we can bypass
the issue and not fail the link check.

[1] https://github.com/apache/camel-kafka-connector/blob/0b4328c9cabe2c1d517f6d042b628be88d15c3c5/docs/modules/ROOT/pages/getting-started.adoc#L10