chore: replace link-checker with cargo-deadlinks
`link-checker` has served us well until recently when it started
reporting false positives. Also, performance is not stellar.

This replaces it with `cargo-deadlinks`[1] which performs much better
(seconds vs minutes) and has almost all functionality of the
`link-checker`. I'm working on bringing it up to feature parity.

Since the `cargo-deadlinks` is a Rust binary, a Yarn plugin was
developed to download the released binary from GitHub releases and makes
it available to Yarn CLI.

Currently a forked version of cargo-deadlinks is used until outstanding
feature work is integrated in the upstream version.

[1] https://github.com/deadlinks/cargo-deadlinks/