chore: add related content to the blog post
Configures and adds a up to five summaries of related blog posts to the
end of a blog post. Also tweaks the summary blog post display by moving
the publishing date to the header, below the title and liks to the blog
post from the title as well from the "Continue reading" link.