refactor: responsive design
With media selectors `max-width: 1023px` and `min-width: 1024px` the
breakpoint was set at 1023px not at 1024px as expected. This increases
`max-width` and `min-width` by 1px.

Also removes the `max-width` of `40rem` so we can occupy the full width
on smaller (<=1024px) resolutions.

And also centers images on the blog.