chore: add custom link as an download option
This adds custom link as an option for downloads. This way we can link
to binary downloads hosted elsewhere, i.e. other than the ASF list, but
still most likely on the camel.apache.org website.

Even though it supports version substitution via `{version}` variable,
for Camel Kafka Connectors it links to the documentation and we only
publish the "latest" version. There is also an issue that a release
version, say `1.2.3` doesn't match up with the documentation version, as
we do it today, i.e. we would have documentation version of `1.2.x`. The
solution might be to publish documentation for each version, but that
will take a toll on the website build times, and we can't really have
docs for all versions ever released, the website would be become really
unmanageable.

So this is a first pass, just to help users find the downloads, and we
need to figure what we would like to do WRT versioning.