fix: nav-explore not showing anything but current
A bug was introduced in fb75497471e8975238ed6114e3550e01f02c533f, in
effort to remove the Camel Quarkus Examples from the nav-explore planel
an JavaScript object was converted to an Array, and when rendered only
the first item was considered. This fixes this issue by making sure that
the structure of the data doesn't change, i.e. given an object it will
output an object when filtering out data for the nav-explore panel.