CAMEL-16576: Move main contributing page from user manual to Community content in camel-website.
The content in community/contributing.md is a combination of CONTRIBUTING.md at the root of the Camel GitHub
repository and the contributing.adoc page in the user manual. The CONTRIBUTING.md file has been shortened and
links to the community/contributing page. The Community section of the User mnanual is removed since all pages
except Release Guide are now in the Community section. The Release Guide and a page with hints for updating the
documentation in the User Manual are now in the "Resources and Guides" section. Links to Contributing in the
footers and a few other pages have been updated.