chore: proxy to pass authentication for GitHub API
Implements and configures a proxy used add `Authorization` header to
invocations of GitHub REST API.

As we're sometimes failing the build due to GitHub API rate limits, and
since in Hugo we can't pass HTTP headers to include the `Authorization`
header which would allow us a higher API invocation limit. An
alternative to use a proxy through which all invocations of the GitHub
API will go through and have the `Authorization` header added.

To use the proxy three environment variables must be set at the same
time:
 - `HUGO_PARAMS_GitHubAPI=http://localhost:22635`
 - `GITHUB_USR=<GitHub username>`
 - `GITHUB_PSW=<GitHub token>`