chore: redesign the release downloads
With this now we have `eol` front matter that'll be added to the new
_Notes_ column in the download table which, replaces the _Description_
column; also features a link to the release notes, previously linked
from the version number. The description of the download is now in
parentheses next to the download link.

The kind of the download (LTS/latest/legacy) is now designed as a label
to make it more prominent.

Also shortens the column headings for better layout on smaller screens
and compacts the PGP and SHA512 columns into one.