chore: build content scheduled for future in PRs
In pull requests checks can't be run unless the HTML files that are
checked are generated. So for files with dates in the future no HTML
files are generated and thus checks might pass as nothing was checked.

This adds `HUGO_OPTIONS` environment variable to `build:hugo` script,
that allows passing `--buildFuture` for pull request builds. This
doesn't have any effect on the production builds as those are done on
ASF Jenkins.