chore: prepare jsonpath pipeline
This is the last bit needed for the changes in

https://github.com/apache/camel/pull/6040

To get merged, utilizing the jsonpath macro in the Antora pipeline to
generate component/endpoint options table.

This has been picked from

https://github.com/apache/camel-website/pull/617

to ease merging. The bit that was not included in this or in the PR #620
is the refactoring around absolute/relative URLs, i.e. the
`patch-sitemap.js` is still present after this.