refactor: latest -> next redirects
This makes sure we redirect from `/<Antora component/latest...` to
`/<Antora component/next...`. A test script was added in
`tests/redirect.sh` so we can test this live or in development; and a
small tweak was done to the way Apache HTTPD is run for development -
for some reason it was terminating for me with SIGWINCH (window resize).