fix: don't rebuild dependencies on Jenkins
This adds the correct dependencies to be used when building on
ci-builds.apache.org Jenkins instance and removes the rebuilding of
dependencies.

We do not want to build or download any dependencies outside of the
already established set of dependencies cached in Yarn cache, this leads
to slow and unstable builds.

The build on Jenkins is performed within a stable environment provided
by the `Dockerfile` and this set of dependencies has been created by
running Yarn within a container created from that.

Fixes #660