feat: cache bursting for UI images
This configures the UI build to generate files with the hash within them
for images in addition to the CSS and JavaScript files we had. Also
layouts/partials that were pointing to the un-hashed versions were
updated.

Since images are quite possibly linked from other websites, most likely
the logo image, a set of redirect rules was added not to break those
sites.

An opportunity to cleanup some long standing CSS issues was also taken.

Fixes #702