chore: additional check that website is up
When running the web server in a container for the redirect tests there
could be some network issues reaching it. This adds detection of that
situation and switches to using curl from inside of the web server
container.

This issue has occurred once on ASF Jenkins in build 622[1],
unfortunately without details allowing us to troubleshot this issue.

[1] https://ci-builds.apache.org/job/Camel/job/Camel.website/job/main/622/