chore: migrate to the new Algolia search app
We have received invitation from Algolia to use the full Algolia Crawler
platform that was previously reserved for paid customers and now is
available for free to DocSearch projects as well. The Algolia Crawler
will replace the current DocSearch platform in three months. We were
given new Application ID and Search API Keys to move over to the new
setup.