refactor(search): align with new index
The Algolia Crawler configuration is now under our direct control, so
we can extract attributes that we could not extract with the DocSearch
crawler. We can now set the `version` attribute, and read it from there
rather than from `lvl6`. There is also `categories` attribute, which
currently we do not display in the search results.
Some refinement and simplifications are also included.