fix: search all components in component search
The `display: none` would remain for non-expanded groups of components.
So when filtering they would not be shown even if matched. This changes
the approach of toggling `display` via JavaScript to toggling it via
CSS. This way we don't need to keep state of non-expanded and then
expanded for the search to show them items. But rather the state pivots
around any text put in the `input` element used for the search. This
makes state handling much simpler and lets the cascading property of CSS
do all the work.

Fixes #789