Remove newline after the link
We need to add custom comment to remove the newline at the end of the
file in the `render-link.html` so that there is no newline after the
rendered link, leading to an additional whitespace before the next
character. Most commonly this results in a " ." instead of "." following
the link.