Fix security list year table row heading
The year list started showing 2023, this was because a security issue
was reported for the year 2022 and previously there weren't any. The
sequence to generate the year in the list was generated from the current
year to the last reported issue's year plus one. This _kind of_ worked
when there was no reports in the current year.

Now the list of years is generated from the current year to the earliest
year with a reported issue, and the security pages are searched to
determine if there are any issues for that year. From that information
it can be determined if the "No issues reported" should be shown or not.