feat: add SBOM downloads based on cut-off date
With this the SBOM downloads will be available for any release created
after 2023-10-11. Both JSON and XML download links will be available.
The Apache conventions for distributing these via dist.apache.org must
be followed, so .asc and .sha512 files must be present in addition to
the .json and .xml SBOM files.

Resolves #1096